Eliminate <jena-core>/jena.cmdline, deprecate rdfcat
