+++ jena-core/src/main/java/jena/rdfcat.java
@@ -24,14 +24,12 @@ package jena;
 // Imports
 ///////////////
 
-import static jena.cmdline.CmdLineUtils.setLog4jConfiguration ;
+import static org.apache.jena.atlas.logging.LogCtl.setCmdLogging;
 
 import java.io.OutputStream ;
 import java.util.* ;
+import java.util.function.BiConsumer;
 
-import jena.cmdline.ArgDecl ;
-import jena.cmdline.ArgHandler ;
-import jena.cmdline.CommandLine ;
 import org.apache.jena.rdf.model.* ;
 import org.apache.jena.rdf.model.impl.IO_Ctl ;
 import org.apache.jena.rdf.model.impl.RDFWriterFImpl ;
@@ -40,8 +38,6 @@ import org.apache.jena.util.FileManager ;
 import org.apache.jena.util.FileUtils ;
 import org.apache.jena.vocabulary.OWL ;
 import org.apache.jena.vocabulary.RDFS ;
-import org.slf4j.Logger ;
-import org.slf4j.LoggerFactory ;
 
 
 /**
@@ -122,112 +118,11 @@ import org.slf4j.LoggerFactory ;
  * input documents, and does not preserve any statement ordering from the input
  * serialisations. Also, duplicate triples will be suppressed.</p>
  */
+@Deprecated
 public class rdfcat
 {
-    static { setLog4jConfiguration() ; }
+    static { setCmdLogging("jena-log4j.properties") ; }
     
-    // Constants
-    //////////////////////////////////
-
-    /** Argument setting expected input language to N3 */
-    public final ArgDecl IN_N3 = new ArgDecl( true, "n", "n3", "ttl", "N3",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    m_actionQ.add( new ReadAction( val, "N3") );
-            }} );
-
-    /** Argument setting expected input language to RDF/XML */
-    public final ArgDecl IN_RDF_XML = new ArgDecl( true, "x", "xml", "rdfxml", "rdf",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    m_actionQ.add( new ReadAction( val, "RDF/XML") );
-            }} );
-
-    /** Argument setting expected input language to NTRIPLE */
-    public final ArgDecl IN_NTRIPLE = new ArgDecl( true, "t", "ntriples", "ntriple", "n-triple", "n-triples",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    m_actionQ.add( new ReadAction( val, "N-TRIPLE" ) );
-            }} );
-
-    /** Argument to set the output language */
-    public final ArgDecl OUT_LANG = new ArgDecl( true, "out",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    setOutput( val );
-            }} );
-
-    /** Argument to set the default input language */
-    public final ArgDecl IN_LANG = new ArgDecl( true, "in",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    expectInput( val );
-            }} );
-
-    /** Argument to turn include processing on */
-    public final ArgDecl INCLUDE = new ArgDecl( false, "include",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    setInclude( true );
-            }} );
-
-    /** Argument to turn include processing off */
-    public final ArgDecl NOINCLUDE = new ArgDecl( false, "noinclude",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    setInclude( false );
-            }} );
-
-    /** Argument to leave import/seeAlso statements in place in flattened models */
-    public final ArgDecl NOFILTER = new ArgDecl( false, "nofilter",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    setRemoveIncludeStatements( false );
-            }} );
-
-    /** Argument to show usage */
-    public final ArgDecl HELP = new ArgDecl( false, "help",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    usage();
-            }} );
-    public final ArgDecl USAGE = new ArgDecl( false, "usage",
-            new ArgHandler() {
-                @Override
-                public void action( String arg, String val ) throws IllegalArgumentException {
-                    usage();
-            }} );
-
-
-    // Static variables
-    //////////////////////////////////
-
-    static private Logger log = LoggerFactory.getLogger( rdfcat.class );
-
-    // Instance variables
-    //////////////////////////////////
-
-    /** The command line processor that handles the arguments */
-    protected CommandLine m_cmdLine = new RCCommandLine().add( IN_N3 )
-                                                         .add( IN_NTRIPLE )
-                                                         .add( IN_RDF_XML )
-                                                         .add( OUT_LANG )
-                                                         .add( IN_LANG )
-                                                         .add( INCLUDE )
-                                                         .add( NOINCLUDE )
-                                                         .add( NOFILTER )
-                                                         .add( HELP )
-                                                         .add( USAGE );
-
     /** The merged model containing all of the inputs */
     protected Model m_model = ModelFactory.createDefaultModel();
 
@@ -249,6 +144,55 @@ public class rdfcat
     /** Action queue */
     protected List<RCAction> m_actionQ = new ArrayList<>();
 
+    
+    // Constants
+    //////////////////////////////////
+
+    /** Argument setting expected input language to N3 */
+    public final ArgDecl IN_N3 = new ArgDecl( true, "n", "n3", "ttl", "N3",
+    		(arg,val) -> m_actionQ.add( new ReadAction( val, "N3") ) );
+
+    /** Argument setting expected input language to RDF/XML */
+    public final ArgDecl IN_RDF_XML = new ArgDecl( true, "x", "xml", "rdfxml", "rdf",
+    		(arg,val) -> m_actionQ.add( new ReadAction( val, "RDF/XML") ) );
+
+    /** Argument setting expected input language to NTRIPLE */
+    public final ArgDecl IN_NTRIPLE = new ArgDecl( true, "t", "ntriples", "ntriple", "n-triple", "n-triples",
+    		(arg,val) -> m_actionQ.add( new ReadAction( val, "N-TRIPLE" ) ) );    
+
+    /** Argument to set the output language */
+    public final ArgDecl OUT_LANG = new ArgDecl( true, "out", (arg,val) -> setOutput( val ) );
+
+    /** Argument to set the default input language */
+    public final ArgDecl IN_LANG = new ArgDecl( true, "in", (arg,val) -> expectInput( val ) );
+
+    /** Argument to turn include processing on */
+    public final ArgDecl INCLUDE = new ArgDecl( false, "include", (arg,val) -> setInclude( true ) );
+
+    /** Argument to turn include processing off */
+    public final ArgDecl NOINCLUDE = new ArgDecl( false, "noinclude", (arg,val) -> setInclude( false ) );
+
+    /** Argument to leave import/seeAlso statements in place in flattened models */
+    public final ArgDecl NOFILTER = new ArgDecl( false, "nofilter", (arg,val) -> setRemoveIncludeStatements( false ) );
+
+    /** Argument to show usage */
+    public final ArgDecl HELP = new ArgDecl( false, "help", (arg,val) -> usage() );
+    public final ArgDecl USAGE = new ArgDecl( false, "usage", (arg,val) -> usage() );
+
+    // Instance variables
+    //////////////////////////////////
+
+    /** The command line processor that handles the arguments */
+    protected CommandLine m_cmdLine = new RCCommandLine().add( IN_N3 )
+                                                         .add( IN_NTRIPLE )
+                                                         .add( IN_RDF_XML )
+                                                         .add( OUT_LANG )
+                                                         .add( IN_LANG )
+                                                         .add( INCLUDE )
+                                                         .add( NOINCLUDE )
+                                                         .add( NOFILTER )
+                                                         .add( HELP )
+                                                         .add( USAGE );
 
     // Constructors
     //////////////////////////////////
@@ -265,6 +209,11 @@ public class rdfcat
 
     /* main loop */
     protected void go( String[] args ) {
+    		
+    		System.err.println( "------------------------------------" );
+		System.err.println( "DEPRECATED: Please use riot instead." );
+		System.err.println( "------------------------------------\n" );
+    	
         // ensure we use the new RIOT parser subsystem
         enableRIOTParser();
 
@@ -418,6 +367,9 @@ public class rdfcat
     }
 
     protected void usage() {
+    		System.err.println( "------------------------------------" );
+    		System.err.println( "DEPRECATED: Please use riot instead." );
+    		System.err.println( "------------------------------------\n" );
         System.err.println( "Usage: java jena.rdfcat (option|input)*" );
         System.err.println( "Concatenates the contents of zero or more input RDF documents." );
         System.err.println( "Options: -out N3 | N-TRIPLE | RDF/XML | RDF/XML-ABBREV" );
@@ -453,6 +405,7 @@ public class rdfcat
     //==============================================================================
 
     /** Local extension to CommandLine to handle mixed arguments and values */
+    @SuppressWarnings("deprecation")
     protected class RCCommandLine
         extends CommandLine
     {
@@ -504,6 +457,7 @@ public class rdfcat
     }
 
     /** Action to set the output format */
+    @SuppressWarnings("deprecation")
     protected class ReadAction
         implements RCAction
     {
@@ -528,4 +482,675 @@ public class rdfcat
             rc.expectInput( l );
         }
     }
+    
+    /**
+     * Command line argument processing based on a trigger model.
+     * An action is called whenever an argument is encountered. Example:
+     * <CODE>
+     * public static void main (String[] args)
+     * {
+     *  CommandLine cl = new CommandLine() ;
+     *  cl.add(false, "verbose")
+     *    .add(true, "--file") ;
+     *  cl.process(args) ;
+     *
+     *  for ( Iterator iter = cl.args() ; iter.hasNext() ; )
+     *  ...
+     * }
+     * </CODE>
+     * A gloabl hook is provided to inspect arguments just before the
+     * action.  Tracing is enabled by setting this to a suitable function
+     * such as that provided by trace():
+     * <CODE>
+     *  cl.setHook(cl.trace()) ;
+     * </CODE>
+     *
+     * <ul>
+     * <li>Neutral as to whether options have - or --</li>
+     * <li>Does not allow multiple single letter options to be concatenated.</li>
+     * <li>Options may be ended with - or --</li>
+     * <li>Arguments with values can use "="</li>
+     * </ul>
+     */
+
+
+    static class CommandLine
+    {
+        /* Extra processor called before the registered one when set.
+         * Used for tracing.
+         */
+        protected BiConsumer<String,String> argHook = null ;
+        protected String usage = null ;
+        protected Map<String, ArgDecl> argMap = new HashMap<>() ;
+        protected Map<String, Arg> args = new HashMap<>() ;
+        //protected boolean ignoreUnknown = false ;
+
+        // Rest of the items found on the command line
+        String indirectionMarker = "@" ;
+        protected boolean allowItemIndirect = false ;   // Allow @ to mean contents of file
+        boolean ignoreIndirectionMarker = false ;       // Allow comand line items to have leading @ but strip it.
+        protected List<String> items = new ArrayList<>() ;
+
+
+        /** Creates new CommandLine */
+        public CommandLine()
+        {
+        }
+
+        /** Set the global argument handler.  Called on every valid argument.
+         * @param argHandler Handler
+         */
+        public void setHook(BiConsumer<String, String> argHandler) { argHook = argHandler ; }
+
+        public void setUsage(String usageMessage) { usage = usageMessage ; }
+
+        public boolean hasArgs() { return args.size() > 0 ; }
+        public boolean hasItems() { return items.size() > 0 ; }
+
+        public Iterator<Arg> args() { return args.values().iterator() ; }
+//        public Map args() { return args ; }
+//        public List items() { return items ; }
+
+        public int numArgs() { return args.size() ; }
+        public int numItems() { return items.size() ; }
+        public void pushItem(String s) { items.add(s) ; }
+
+        public boolean isIndirectItem(int i)
+        { return allowItemIndirect && items.get(i).startsWith(indirectionMarker) ; }
+
+        public String getItem(int i)
+        {
+            return getItem(i, allowItemIndirect) ;
+        }
+
+        public String getItem(int i, boolean withIndirect)
+        {
+            if ( i < 0 || i >= items.size() )
+                return null ;
+
+
+            String item = items.get(i) ;
+
+            if ( withIndirect && item.startsWith(indirectionMarker) )
+            {
+                item = item.substring(1) ;
+                try { item = FileUtils.readWholeFileAsUTF8(item) ; }
+                catch (Exception ex)
+                { throw new IllegalArgumentException("Failed to read '"+item+"': "+ex.getMessage()) ; }
+            }
+            return item ;
+        }
+
+
+        /** Process a set of command line arguments.
+         * @param argv The words of the command line.
+         * @throws IllegalArgumentException Throw when something is wrong (no value found, action fails).
+         */
+        public void process(String[] argv) throws java.lang.IllegalArgumentException
+        {
+            List<String> argList = new ArrayList<>() ;
+            argList.addAll(Arrays.asList(argv)) ;
+
+            int i = 0 ;
+            for ( ; i < argList.size() ; i++ )
+            {
+                String argStr = argList.get(i) ;
+                if (endProcessing(argStr))
+                    break ;
+                
+                if ( ignoreArgument(argStr) )
+                    continue ;
+
+                // If the flag has a "=" or :, it is long form --arg=value.
+                // Split and insert the arg
+                int j1 = argStr.indexOf('=') ;
+                int j2 = argStr.indexOf(':') ;
+                int j = Integer.MAX_VALUE ;
+
+                if ( j1 > 0 && j1 < j )
+                    j = j1 ;
+                if ( j2 > 0 && j2 < j )
+                    j = j2 ;
+
+                if ( j != Integer.MAX_VALUE )
+                {
+                    String a2 = argStr.substring(j+1) ;
+                    argList.add(i+1,a2) ;
+                    argStr = argStr.substring(0,j) ;
+                }
+
+                argStr = ArgDecl.canonicalForm(argStr) ;
+                String val = null ;
+
+                if ( argMap.containsKey(argStr) )
+                {
+                    if ( ! args.containsKey(argStr))
+                        args.put(argStr, new Arg(argStr)) ;
+
+                    Arg arg = args.get(argStr) ;
+                    ArgDecl argDecl = argMap.get(argStr) ;
+
+                    if ( argDecl.takesValue() )
+                    {
+                        if ( i == (argList.size()-1) )
+                            throw new IllegalArgumentException("No value for argument: "+arg.getName()) ;
+                        i++ ;
+                        val = argList.get(i) ;
+                        arg.setValue(val) ;
+                        arg.addValue(val) ;
+                    }
+
+                    // Global hook
+                    if ( argHook != null )
+                        argHook.accept(argStr, val) ;
+
+                    argDecl.trigger(arg) ;
+                }
+                else
+                    handleUnrecognizedArg( argList.get(i) );
+//                    if ( ! getIgnoreUnknown() )
+//                        // Not recognized
+//                        throw new IllegalArgumentException("Unknown argument: "+argStr) ;
+            }
+
+            // Remainder.
+            if ( i < argList.size() )
+            {
+                if ( argList.get(i).equals("-") || argList.get(i).equals("--") )
+                    i++ ;
+                for ( ; i < argList.size() ; i++ )
+                {
+                    String item = argList.get(i) ;
+                    items.add(item) ;
+                }
+            }
+        }
+
+        /** Hook to test whether this argument should be processed further
+         */
+        public boolean ignoreArgument( String argStr )
+        { return false ; }
+        
+        /** Answer true if this argument terminates argument processing for the rest
+         * of the command line. Default is to stop just before the first arg that
+         * does not start with "-", or is "-" or "--".
+         */
+        public boolean endProcessing( String argStr )
+        {
+            return ! argStr.startsWith("-") || argStr.equals("--") || argStr.equals("-");
+        }
+
+        /**
+         * Handle an unrecognised argument; default is to throw an exception
+         * @param argStr The string image of the unrecognised argument
+         */
+        public void handleUnrecognizedArg( String argStr ) {
+            throw new IllegalArgumentException("Unknown argument: "+argStr) ;
+        }
+
+
+        /** Test whether an argument was seen.
+         */
+
+        public boolean contains(ArgDecl argDecl) { return getArg(argDecl) != null ; }
+
+        /** Test whether an argument was seen.
+         */
+
+        public boolean contains(String s) { return getArg(s) != null ; }
+
+
+        /** Test whether the command line had a particular argument
+         *
+         * @param argName
+         */
+        public boolean hasArg(String argName) { return getArg(argName) != null ; }
+
+        /** Test whether the command line had a particular argument
+         *
+         * @param argDecl
+         */
+
+        public boolean hasArg(ArgDecl argDecl) { return getArg(argDecl) != null ; }
+
+
+        /** Get the argument associated with the argument declaration.
+         *  Actually returns the LAST one seen
+         *  @param argDecl Argument declaration to find
+         *  @return Last argument that matched.
+         */
+
+        public Arg getArg(ArgDecl argDecl)
+        {
+            Arg arg = null ;
+            for ( Arg a : args.values() )
+            {
+                if ( argDecl.matches( a ) )
+                {
+                    arg = a;
+                }
+            }
+            return arg ;
+        }
+
+        /** Get the argument associated with the arguement name.
+         *  Actually returns the LAST one seen
+         *  @param arg Argument declaration to find
+         *  @return Arg - Last argument that matched.
+         */
+
+        public Arg getArg(String arg)
+        {
+            arg = ArgDecl.canonicalForm(arg) ;
+            return args.get(arg) ;
+        }
+
+        /**
+         * Returns the value (a string) for an argument with a value -
+         * returns null for no argument and no value.
+         * @param argDecl
+         * @return String
+         */
+        public String getValue(ArgDecl argDecl)
+        {
+            Arg arg = getArg(argDecl) ;
+            if ( arg == null )
+                return null ;
+            if ( arg.hasValue())
+                return arg.getValue() ;
+            return null ;
+        }
+
+        /**
+         * Returns the value (a string) for an argument with a value -
+         * returns null for no argument and no value.
+         * @param argName
+         * @return String
+         */
+        public String getValue(String argName)
+        {
+            Arg arg = getArg(argName) ;
+            if ( arg == null )
+                return null ;
+            return arg.getValue() ;
+        }
+
+        /**
+         * Returns all the values (0 or more strings) for an argument.
+         * @param argDecl
+         * @return List
+         */
+        public List<String> getValues(ArgDecl argDecl)
+        {
+            Arg arg = getArg(argDecl) ;
+            if ( arg == null )
+                return null ;
+            return arg.getValues() ;
+        }
+
+        /**
+         * Returns all the values (0 or more strings) for an argument.
+         * @param argName
+         * @return List
+         */
+        public List<String> getValues(String argName)
+        {
+            Arg arg = getArg(argName) ;
+            if ( arg == null )
+                return null ;
+            return arg.getValues() ;
+        }
+
+
+
+        /** Add an argument to those to be accepted on the command line.
+         * @param argName Name
+         * @param hasValue True if the command takes a (string) value
+         * @return The CommandLine processor object
+         */
+
+        public CommandLine add(String argName, boolean hasValue)
+        {
+            return add(new ArgDecl(hasValue, argName)) ;
+        }
+
+        /** Add an argument to those to be accepted on the command line.
+         *  Argument order reflects ArgDecl.
+         * @param hasValue True if the command takes a (string) value
+         * @param argName Name
+         * @return The CommandLine processor object
+         */
+
+        public CommandLine add(boolean hasValue, String argName)
+        {
+            return add(new ArgDecl(hasValue, argName)) ;
+        }
+
+        /** Add an argument object
+         * @param arg Argument to add
+         * @return The CommandLine processor object
+         */
+
+        public CommandLine add(ArgDecl arg)
+        {
+            for ( Iterator<String> iter = arg.names() ; iter.hasNext() ; )
+                argMap.put(iter.next(), arg) ;
+            return this ;
+        }
+
+//        public boolean getIgnoreUnknown() { return ignoreUnknown ; }
+//        public void setIgnoreUnknown(boolean ign) { ignoreUnknown = ign ; }
+
+        /**
+         * @return Returns whether items starting "@" have the value of named file.
+         */
+        public boolean allowItemIndirect()
+        {
+            return allowItemIndirect ;
+        }
+
+        /**
+         * @param allowItemIndirect Set whether items starting "@" have the value of named file.
+
+         */
+        public void setAllowItemIndirect(boolean allowItemIndirect)
+        {
+            this.allowItemIndirect = allowItemIndirect ;
+        }
+
+        /**
+         * @return Returns the ignoreIndirectionMarker.
+         */
+        public boolean isIgnoreIndirectionMarker()
+        {
+            return ignoreIndirectionMarker ;
+        }
+
+        /**
+         * @return Returns the indirectionMarker.
+         */
+        public String getIndirectionMarker()
+        {
+            return indirectionMarker ;
+        }
+
+        /**
+         * @param indirectionMarker The indirectionMarker to set.
+         */
+        public void setIndirectionMarker(String indirectionMarker)
+        {
+            this.indirectionMarker = indirectionMarker ;
+        }
+
+        /**
+         * @param ignoreIndirectionMarker The ignoreIndirectionMarker to set.
+         */
+        public void setIgnoreIndirectionMarker(boolean ignoreIndirectionMarker)
+        {
+            this.ignoreIndirectionMarker = ignoreIndirectionMarker ;
+        }
+
+	    	public BiConsumer<String, String> trace() {
+	    		return (arg, val) -> {
+	    			System.err.println("Seen: " + arg + (val != null ? " = " + val : ""));
+	    		};
+	    	}
+
+    }
+    
+    /** A command line argument that has been foundspecification.
+	 */
+	static class Arg
+	{
+	    String name ;
+	    String value ;
+	    List<String> values = new ArrayList<>() ;
+	    
+	    Arg() { name = null ; value = null ; }
+	    
+	    Arg(String _name) { this() ; setName(_name) ; }
+	    
+	    Arg(String _name, String _value) { this() ; setName(_name) ; setValue(_value) ; }
+	    
+	    void setName(String n) { name = n ; }
+	    
+	    void setValue(String v) { value = v ; }
+	    void addValue(String v) { values.add(v) ; }
+	    
+	    public String getName() { return name ; }
+	    public String getValue() { return value; }
+	    public List<String> getValues() { return values; }
+	    
+	    public boolean hasValue() { return value != null ; }
+	    
+	    public boolean matches(ArgDecl decl)
+	    {
+	        return decl.getNames().contains(name) ;
+	    }
+	    
+	}
+    
+    /** A command line argument specification.
+	 */
+	static class ArgDecl
+	{
+	    boolean takesValue ;
+	    Set<String> names = new HashSet<>() ;
+	    boolean takesArg = false ;
+		List<BiConsumer<String, String>> argHooks = new ArrayList<>() ;
+	    public static final boolean HasValue = true ;
+	    public static final boolean NoValue = false ;
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     */
+
+	    public ArgDecl(boolean hasValue)
+	    {
+	        takesValue = hasValue ;
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name      Name of argument
+	     */
+
+	    public ArgDecl(boolean hasValue, String name)
+	    {
+	        this(hasValue) ;
+	        addName(name) ;
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name      Name of argument
+	     * @param handler   BiConsumer<String, String>
+	     */
+
+	    public ArgDecl(boolean hasValue, String name, BiConsumer<String, String> handler)
+	    {
+	        this(hasValue) ;
+	        addName(name) ;
+	        addHook( handler );
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param handler   BiConsumer<String, String>
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, BiConsumer<String, String> handler)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addHook( handler );
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param name3      Name of argument
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, String name3)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addName(name3) ;
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param name3      Name of argument
+	     * @param handler   BiConsumer<String, String>
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, String name3, BiConsumer<String, String> handler)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addName(name3) ;
+	        addHook( handler );
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param name3      Name of argument
+	     * @param name4      Name of argument
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, String name3, String name4)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addName(name3) ;
+	        addName(name4) ;
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param name3      Name of argument
+	     * @param name4      Name of argument
+	     * @param handler    BiConsumer<String, String>
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, String name3, String name4, BiConsumer<String, String> handler)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addName(name3) ;
+	        addName(name4) ;
+	        addHook( handler );
+	    }
+
+	    /** Create a declaration for a command argument.
+	     *
+	     * @param hasValue  Does it take a value or not?
+	     * @param name1      Name of argument
+	     * @param name2      Name of argument
+	     * @param name3      Name of argument
+	     * @param name4      Name of argument
+	     * @param name5      Name of argument
+	     * @param handler    BiConsumer<String, String>
+	     */
+
+	    public ArgDecl(boolean hasValue, String name1, String name2, String name3, String name4, String name5, BiConsumer<String, String> handler)
+	    {
+	        this(hasValue) ;
+	        addName(name1) ;
+	        addName(name2) ;
+	        addName(name3) ;
+	        addName(name4) ;
+	        addName(name5) ;
+	        addHook( handler );
+	    }
+
+	    public void addName(String name)
+	    {
+	        name = canonicalForm(name) ;
+	        names.add(name) ;
+	    }
+
+	    public Set<String> getNames() { return names ; }
+	    public Iterator<String> names() { return names.iterator() ; }
+
+	    // Callback model
+
+	    public void addHook(BiConsumer<String, String> argHandler)
+	    {
+	        argHooks.add(argHandler) ;
+	    }
+
+	    protected void trigger(Arg arg)
+	    {
+			argHooks.forEach(action -> action.accept( arg.getName(), arg.getValue() ));
+	    }
+
+	    public boolean takesValue() { return takesValue ; }
+
+	    public boolean matches(Arg a)
+	    {
+	        for ( String n : names )
+	        {
+	            if ( a.getName().equals( n ) )
+	            {
+	                return true;
+	            }
+	        }
+	        return false ;
+	    }
+
+	    public boolean matches(String arg)
+	    {
+	        arg = canonicalForm(arg) ;
+	        return names.contains(arg) ;
+	    }
+
+	    static String canonicalForm(String str)
+	    {
+	        if ( str.startsWith("--") )
+	            return str.substring(2) ;
+
+	        if ( str.startsWith("-") )
+	            return str.substring(1) ;
+
+	        return str ;
+	    }
+	}
 }
