+++ jena-core/src/main/java/jena/rdfcompare.java
@@ -18,7 +18,7 @@
 
 package jena;
 
-import static jena.cmdline.CmdLineUtils.setLog4jConfiguration;
+import static org.apache.jena.atlas.logging.LogCtl.setCmdLogging;
 
 import java.net.URL;
 import java.io.FileInputStream;
@@ -46,7 +46,7 @@ import org.apache.jena.rdf.model.* ;
  */
 public class rdfcompare extends java.lang.Object {
 
-    static { setLog4jConfiguration() ; }
+    static { setCmdLogging("jena-log4j.properties") ; }
 
     /**
     * @param args the command line arguments
