+++ jena-core/src/main/java/jena/rset.java
@@ -18,13 +18,12 @@
 
 package jena;
 
-import jena.cmdline.CmdLineUtils ;
 
 public class rset
 {
     // Call-through to arq command line application
     public static void main(String... args)
     {
-        CmdLineUtils.invokeCmd("arq.rset", args) ;
+        InvokingUtil.invokeCmd("arq.rset", args) ;
     }
 }
