+++ jena-core/src/main/java/jena/sparql.java
@@ -18,13 +18,12 @@
 
 package jena;
 
-import jena.cmdline.CmdLineUtils ;
 
 public class sparql
 {
     // Call-through to arq command line application
     public static void main(String[] args)
     {
-        CmdLineUtils.invokeCmd("arq.sparql",args) ;
+        InvokingUtil.invokeCmd("arq.sparql",args) ;
     }
 }
