+++ jena-core/src/main/java/jena/version.java
@@ -18,7 +18,7 @@
 
 package jena;
 
-import static jena.cmdline.CmdLineUtils.setLog4jConfiguration;
+import static org.apache.jena.atlas.logging.LogCtl.setCmdLogging;
 
 import java.lang.reflect.*;
 
@@ -30,7 +30,7 @@ import org.apache.jena.Jena ;
  */
 public class version implements Jena {
 
-    static { setLog4jConfiguration() ; }
+    static { setCmdLogging("jena-log4j.properties") ; }
 
     /**
 	 * Print out jena version information and exit.
